refactor: user domain의 DTO 패키지 분리 및 DTO명 변경 #20
