+++ src/main/java/com/munecting/api/domain/user/dto/request/LoginRequestDto.java
@@ -1,11 +1,11 @@
-package com.munecting.api.domain.user.dto;
+package com.munecting.api.domain.user.dto.request;
 
 import com.munecting.api.domain.user.constant.SocialType;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record UserLoginRequestDto(
+public record LoginRequestDto(
         @NotNull
         @Schema(description = "Identity Provider")
         SocialType socialType,
