+++ src/main/java/com/munecting/api/domain/user/dto/request/LogoutRequestDto.java
@@ -1,0 +1,8 @@
+package com.munecting.api.domain.user.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LogoutRequestDto(
+        @NotBlank String accessToken
+) {
+}
