+++ src/main/java/com/munecting/api/domain/user/dto/response/UserTokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.munecting.api.domain.user.dto;
+package com.munecting.api.domain.user.dto.response;
 
 import lombok.Builder;
